pip_requirements.txt
absl-py==2.3.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
ansicolors==1.1.8
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
beautifulsoup4==4.14.2
Bottleneck @ file:///croot/bottleneck_1731058641041/work
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.3.1
cmdstanpy==1.3.0
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.14
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
deepspeed==0.16.7
diffusers==0.33.1
dill==0.4.0
einops==0.8.1
entrypoints==0.4
et_xmlfile @ file:///croot/et_xmlfile_1756462486635/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.8.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
hjson==3.1.0
holidays==0.83
huggingface-hub==0.30.2
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
ipdb==0.13.13
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1760459840/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1759151815/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib @ file:///work/perseverance-python-buildout/croot/joblib_1728392999985/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver==1.4.8
kloppy==3.18.0
lightgbm==4.6.0
lxml==6.0.0
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1741006473187/work
mplsoccer==1.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.0
nbclient==0.10.2
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
ninja==1.11.1.4
numexpr @ file:///io/perseverance-python-buildout/croot/numexpr_1731701292781/work
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl @ file:///croot/openpyxl_1736366913861/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp313-cp313-linux_x86_64.whl#sha256=248baa16b8f7b48e3a9c0e641fba9a72bce9c95c0dddf7c9a7547016b61fe3fb
pandera==0.22.1
papermill==2.6.0
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
patsy==1.0.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
plotly==6.3.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.0
psutil==7.0.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.3
python-dateutil @ file:///work/perseverance-python-buildout/croot/python-dateutil_1728385355105/work
pytz @ file:///work/perseverance-python-buildout/croot/pytz_1728386814080/work
PyYAML==6.0.2
pyzmq @ file:///home/task_175810349669950/conda-bld/pyzmq_1758103933718/work
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rpds-py==0.29.0
safetensors==0.5.3
salesforce-merlion==2.0.4
scikit-base==0.12.6
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1743153247252/work/dist/scipy-1.15.2-cp313-cp313-linux_x86_64.whl#sha256=0fcdea24ba589efe1517ae1147ed7b83d8b1fcc30dbb90fe961c583a7d975654
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==72.1.0
six @ file:///croot/six_1744271502820/work
sktime==0.39.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
stanio==0.5.1
statsmodels==0.14.5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl @ file:///croot/threadpoolctl_1731006087558/work
tokenizers==0.21.1
torch==2.6.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1756854903354/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.3
triton==3.2.0
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3==2.4.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
Werkzeug==3.1.3
wheel==0.45.1
xlrd==2.0.2
xlutils==2.0.0
xlwt==1.3.0
yarl==1.22.0
zipp==3.21.0
